(#3) Started refactor of dungeon generator using DI.
