+++ Assets/Scripts/Scripts.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dc3a5bcb300761b46ab4e625dda2c6a3
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
